Refs #35255 - split Tablerow component
